Enable dependabot for updating the `github.com/gardener/gardener` dependency (#86)
* Use the upstream golang image in the TestDefinition

In https://github.com/gardener/gardener/pull/7553 we adapted all TestDefinitions in gardener/gardener to use the upstream golang image. The copied GCR image is no longer maintained.

* Add `apiVersion` to the TestDefinition

See https://github.com/gardener/gardener/pull/8293. This is required to make dependabot update the golang image version in the TestDefinition.

* Enable dependabot for updating the `github.com/gardener/gardener` dependency

The commit also enables dependabot to update the golang and alpine images in the Dockerfile and in the TestDefinition.