support gvisor for containerd2.1 with config version=3
gardener-extension-gvisor uses the install-gvisor-containerd.sh script inside a
configmap-containerd.yaml to change the /etc/containerd/config.toml
to make containerd aware of the gvisor installation (aka configure it).

Since Garden Linux 1877 uses containerd 2.1, and containerd 2.1 uses the
format version=3 of the config file, some properties have changed.

This PR makes the required changes to also work with containerd 2.1 (aka config format version=3)

This commit changes the install-gvisor-containerd.sh in a way,
that the new config properties are used so that gvisor can be used with containerd2.1 (i.e. with Garden Linux 1877)

In version=2 the property to define additional runtimes is called:
   [plugins."io.containerd.grpc.v1.cri".containerd.runtimes]

In version=3 the property to define additional runtimes is called:
   [plugins."io.containerd.cri.v1.runtime".containerd.runtimes]